services/knowledge-service/requirements.txt
# Web Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# Database
psycopg2-binary==2.9.9
sqlalchemy==2.0.23

# Vector Database
qdrant-client==1.7.0

# Document Processing
PyPDF2==3.0.1
python-docx==1.1.0
openpyxl==3.1.2
python-magic==0.4.27

# AI/ML
openai==1.40.0
httpx==0.27.0
langchain==0.0.350
tiktoken==0.5.2

# Utilities
python-dotenv==1.0.0
pydantic==2.5.2
pydantic-settings==2.1.0

# Cloud Storage (for future GCP integration)
# google-cloud-storage==2.10.0
